refactor: 공통 헬퍼로직 유틸 디렉토리 분리, 이관작업
